Moving WebClient creation logic to classfixture.
